cms-admin: Rename action log components/header to components/actionLogHeader
Match the sibling folder naming convention (camelCase folder = component
file name) — all other folders under components/ already follow it
(actionChip, actionLogCompare, actionLogShowVersion, diffHeader,
diffViewer, userCell).
